Move some of Texture2DExtensions into new ImageEdit class
Some of these methods are moved to avoid a weird dependency hierarchy when I add assemblies, and some is because it makes more sense to group them with the other moved methods
