+++ Assets/Scripts/Layers/TileLayer.cs
@@ -332,7 +332,7 @@ namespace PAC.Layers
                 tile.bottomLeft += new IntVector2(left, down);
             }
 
-            Texture2DExtensions.ExtendCropOptions extendCropOptions = new Texture2DExtensions.ExtendCropOptions { left = left, right = right, top = up, bottom = down };
+            ImageEdit.ExtendCropOptions extendCropOptions = new ImageEdit.ExtendCropOptions { left = left, right = right, top = up, bottom = down };
             foreach (AnimationKeyFrame keyFrame in keyFrames)
             {
                 keyFrame.texture = keyFrame.texture.ExtendCrop(extendCropOptions);
